@@ -35,11 +35,11 @@
     "node-sass": "^4.11.0",
     "nyc": "^13.1.0",
     "prop-types": "^15.6.2",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
+    "react": "^16.8.4",
+    "react-dom": "^16.8.4",
     "react-hot-loader": "^4.6.3",
     "react-redux": "^6.0.0",
-    "react-test-renderer": "^16.7.0",
+    "react-test-renderer": "^16.8.4",
     "redux": "^4.0.1",
     "sass-loader": "^7.1.0",
     "screenfull": "^4.0.0",
@@ -48,6 +48,7 @@
     "sinon": "^7.2.2",
     "source-map-support": "^0.5.10",
     "style-loader": "^0.23.1",
+    "styled-components": "^4.1.3",
     "toastr": "^2.1.4",
     "webpack": "^4.28.4",
     "webpack-cli": "^3.2.1",
